feat(server/player): "private" playerData and new CPlayer API
Hold additional data in the playerData table. Can be retrieved in other
resources by getting the player reference and calling player.get(index).

Call Ox.CPlayer to get the player object and trigger a method with the
given arguments.

Export a whitelist of player exports to call in other resources, for
use with imports.