feat(server/player): setPlayerInService event
Duty/service was mentioned in Discord. Groups don't have a concept of
going "in service", but it's easily tied to the player with metadata.
Basic and safe implementation; if you don't want it tied to a specific
group (or want multiple groups using the same value) you can handle
it externally.