refactor(server/player): re-structure player module and class
Creates a PlayerRegistry to replace Player.list, and moves GetPlayer functions
to the submodule, along with events related to joining and leaving.

CPlayer is moved into player/class, with several functions and methods
being exchanged or removed.

Annotations still being reworked, but need to sus out the best way to
handle them. Separation of client and server + ease of use are important.
The class defined at the top of player/class is the current WIP.