refactor!: consistent camelCase
As mentioned in #85, we had a mix of snake_case, camelCase, and lowercase.
Though snake_case is the "standard" for SQL, it's easier to match column
names to variables and table fields.

This breaks basically all other resources, including our own, and it has
received little-to-no testing for incorrectly named variables.

Other breaking changes are also incoming for the next release in preparation
for a v1 release - we can't sit in pre-release hell forever.